+++ imod/mf6/wel.py
@@ -50,7 +50,7 @@ class Well(BoundaryCondition):
         "observations",
     )
     _pkg_id = "wel"
-    _binary_data = ("layer", "row", "column", "rate")
+    _period_data = ("layer", "row", "column", "rate")
     _keyword_map = {}
     _template = BoundaryCondition._initialize_template(_pkg_id)